requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aenum=3.1.16=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
attrs=25.3.0=pypi_0
azure-ai-documentintelligence=1.0.2=pypi_0
azure-cognitiveservices-vision-computervision=0.9.1=pypi_0
azure-common=1.1.28=pypi_0
azure-core=1.35.0=pypi_0
boto3=1.38.7=pypi_0
botocore=1.38.7=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
deprecated=1.2.18=pypi_0
expat=2.7.1=h8ddb27b_0
frozenlist=1.6.0=pypi_0
future=1.0.0=pypi_0
idna=3.10=pypi_0
isodate=0.7.2=pypi_0
jmespath=1.0.1=pypi_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
line-bot-sdk=3.17.1=pypi_0
msrest=0.7.1=pypi_0
multidict=6.4.3=pypi_0
oauthlib=3.3.1=pypi_0
openssl=3.0.16=h3f729d1_0
pillow=11.3.0=pypi_0
pip=25.1=pyh0d26453_1
propcache=0.3.1=pypi_0
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
python=3.13.2=hadb2040_100_cp313
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.13=0_cp313
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
s3transfer=0.12.0=pypi_0
six=1.17.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
tk=8.6.14=h0416ee5_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025b=h04d1e81_0
urllib3=2.4.0=pypi_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
wrapt=1.17.2=pypi_0
xz=5.6.4=h4754444_1
yarl=1.20.0=pypi_0
zlib=1.2.13=h8cc25b3_1
